+++ packages/yourloops/test/unit/components/footer/footer.test.tsx
@@ -34,10 +34,14 @@ import i18n from 'i18next'
 import Footer from '../../../../components/footer/footer'
 import { AuthContext, useAuth, User } from '../../../../lib/auth'
 import diabeloopUrls from '../../../../lib/diabeloop-url'
+import { Router } from 'react-router-dom'
+import { createMemoryHistory } from 'history'
+import { triggerMouseEvent } from '../../common/utils'
 
 describe('Footer', () => {
   let auth: AuthContext = null
   let container: HTMLElement | null = null
+  const history = createMemoryHistory({ initialEntries: ['/'] })
 
   const FooterLinksComponent = (data: { user: User }): JSX.Element => {
     auth = useAuth()
@@ -52,7 +56,9 @@ describe('Footer', () => {
     await act(() => {
       return new Promise((resolve) => {
         render(
-          <FooterLinksComponent user={user} />, container, resolve)
+          <Router history={history}>
+            <FooterLinksComponent user={user} />
+          </Router>, container, resolve)
       })
     })
   }
@@ -81,12 +87,10 @@ describe('Footer', () => {
     expect(component).not.toBeNull()
   })
 
-  it('should render language selector and accompanying document selector when user is not logged in', async () => {
+  it('should render language selector when user is not logged in', async () => {
     await mountComponent()
     const languageSelector = document.getElementById('footer-language-box')
-    const documentSelector = document.getElementById('footer-accompanying-documents-box')
     expect(languageSelector).not.toBeNull()
-    expect(documentSelector).not.toBeNull()
   })
 
   it('should not render language selector when user is logged in', async () => {
@@ -100,6 +104,25 @@ describe('Footer', () => {
   it('should privacy policy link redirect to correct url', async () => {
     await mountComponent()
     checkLinkHref('footer-link-url-privacy-policy', diabeloopUrls.getPrivacyPolicyUrL(i18n.language))
+  })
+
+  it('should training link redirect to correct url', async () => {
+    await mountComponent()
+    checkLinkHref('footer-link-url-training', diabeloopUrls.getTrainingUrl(i18n.language))
+  })
+
+  it('should intended use link redirect to /intended-use page when user is logged in', async () => {
+    await mountComponent()
+    const link = document.getElementById('footer-link-intended-use')
+    triggerMouseEvent('click', link)
+    expect(history.location.pathname).toBe('/intended-use')
+  })
+
+  it('should intended use link redirect to /intended-use page when user is not logged in', async () => {
+    await mountComponent({} as User)
+    const link = document.getElementById('footer-link-intended-use')
+    triggerMouseEvent('click', link)
+    expect(history.location.pathname).toBe('/intended-use')
   })
 
   it('should terms of use link redirect to correct url', async () => {